server/pyproject.toml
[tool.poetry]
name = "farm-stack-template-api"
version = "0.1.0"
description = ""
authors = ["Davi Lucciola <davilucciola14@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.112.0"
beanie = "^1.26.0"
uvicorn = "^0.30.6"
httpx = "^0.27.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-decouple = "^3.8"
pydantic-settings = "^2.4.0"
jinja2 = "^3.1.4"
passlib = "^1.7.4"
bcrypt = "4.0.1"
email-validator = "^2.2.0"


[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
pytest = "^8.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
